Move NewChat button path
